fix(resources): resolve circular dependency between resources and posts
* Creates a shared resource-types.ts file to break the circular dependency\n* Reorganizes type definitions to prevent initialization errors\n* Maintains backward compatibility with existing imports
