+++ apps/ai-hero/src/lib/posts.ts
@@ -2,19 +2,13 @@ import { z } from 'zod'
 
 import { ContentResourceSchema } from '@coursebuilder/core/schemas/content-resource-schema'
 
-import { POST_SUBTYPES } from './resources'
+import {
+	POST_SUBTYPES,
+	POST_TYPES_WITH_VIDEO,
+	PostType,
+	PostTypeSchema,
+} from './resource-types'
 import { TagSchema } from './tags'
-
-export const POST_TYPES_WITH_VIDEO = ['article']
-
-export const PostTypeSchema = z.union([
-	z.literal('article'),
-	z.literal('podcast'),
-	z.literal('tip'),
-	z.literal('course'),
-])
-
-export type PostType = z.infer<typeof PostTypeSchema>
 
 export const PostActionSchema = z.union([
 	z.literal('publish'),
@@ -25,11 +19,15 @@ export const PostActionSchema = z.union([
 
 export type PostAction = z.infer<typeof PostActionSchema>
 
+// Re-export for backward compatibility
+export { POST_TYPES_WITH_VIDEO, PostTypeSchema }
+export type { PostType }
+
 export const NewPostInputSchema = z.object({
 	title: z.string().min(1, 'Title is required'),
 	videoResourceId: z.string().optional(),
 	postType: z.string().refine(
-		(type) => POST_SUBTYPES.includes(type),
+		(type) => POST_SUBTYPES.includes(type as PostType),
 		(type) => ({
 			message: `Invalid post type: ${type}. Valid types are: ${POST_SUBTYPES.join(', ')}`,
 		}),
